+++ .github/actions/publish-docker-studio-action/entrypoint.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-docker --version
-source /home/.virtualenvs/py36/bin/activate
-make clean stage deps lint build-docker-studio
-cat logs/packer-docker-studio.log
-echo "${DOCKERHUB_TOKEN}" | docker login --username cliffano --password-stdin
-docker inspect cliffano/studio
-make publish-docker-studio